+++ lib/decorators/src/public_api.ts
@@ -4,3 +4,4 @@ export { StateRepository } from './state-repository/state-repository';
 export { debounce } from './debounce-time/debounce';
 export { payload } from './payload/payload';
 export { named } from './named/named';
+export { computed } from './computed/computed';