+++ lib/typings/src/types/computed-options.ts
@@ -1,0 +1,6 @@
+import { Any } from './any';
+
+export interface ComputedOptions {
+    sequenceId: number;
+    value: Any;
+}